+++ pydantic_encryption/adapters/hashing/__init__.py
@@ -1,0 +1,3 @@
+from pydantic_encryption.adapters.hashing import argon2
+
+__all__ = ["argon2"]
