--- pydantic_encryption/lib/adapters/encryption/__init__.py
@@ -1,3 +1,0 @@
-from . import aws, evervault, fernet
-
-__all__ = ["fernet", "evervault", "aws"]
