--- pydantic_encryption/lib/adapters/hashing/__init__.py
@@ -1,1 +1,0 @@
-from . import argon2
