+++ pydantic_encryption/models/__init__.py
@@ -1,2 +1,3 @@
-from .base import *
-from .secure_model import *
+from pydantic_encryption.models.base import BaseModel, SecureModel
+
+__all__ = ["BaseModel", "SecureModel"]
