+++ pyproject.toml
@@ -14,16 +14,27 @@ requires-python = ">=3.11,<3.15"
 dependencies = [
     "pydantic>=2.10.6",
     "pydantic-settings>=2.9.1",
-    "evervault>=4.4.1",
-    "argon2-cffi>=23.1.0",
-    "aws-encryption-sdk[mpl]>=4.0.1",
-    "boto3>=1.38.8",
     "pydantic-super-model>=1.1.2",
-    "generalimport>=0.5.2",
+    "cryptography>=44.0.0",
+    "argon2-cffi>=23.1.0",
 ]
 
 [project.optional-dependencies]
+aws = [
+    "aws-encryption-sdk[mpl]>=4.0.1",
+    "boto3>=1.38.8",
+]
+evervault = [
+    "evervault>=4.4.1",
+]
 sqlalchemy = [
+    "sqlalchemy>=2.0.40",
+    "sqlmodel>=0.0.24",
+]
+all = [
+    "aws-encryption-sdk[mpl]>=4.0.1",
+    "boto3>=1.38.8",
+    "evervault>=4.4.1",
     "sqlalchemy>=2.0.40",
     "sqlmodel>=0.0.24",
 ]
@@ -36,7 +47,7 @@ dev = [
     "pytest-cov>=6.1.1",
     "pytest-sqlalchemy>=0.3.0",
     "sqlalchemy-utils>=0.41.2",
-    "psycopg2-binary>=2.9.10"
+    "psycopg2-binary>=2.9.10",
 ]
 
 [tool.black]
